Add files via upload
Optimized IForest-based Intrusion Detection System (OIFIDS) is designed to handle both heterogeneous and streaming data. It employs a collection of optimized binary trees, each of which is trained on a distinctive subset of data, and in which the location of empty leaves determines the anomaly score assigned to a certain data point. It utilizes a modified version of the Harris Hawks Optimization algorithm (ERHHO). It is evaluated using three datasets: CICIDS-2018, NSL-KDD, and UNSW-NB15.